erro no avaliador de rrr

+++ srcs/testargs.c
@@ -6,7 +6,7 @@
 /*   By: mde-figu <mde-figu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/27 17:04:26 by mde-figu          #+#    #+#             */
-/*   Updated: 2021/09/14 18:05:39 by mde-figu         ###   ########.fr       */
+/*   Updated: 2021/09/16 15:47:03 by mde-figu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,7 @@ static _Bool	appears_twice(char *c, char **argv, size_t index)
 void	testargs(char **argv)
 {
 	char	**ptr;
-	int		i;
+	size_t		i;
 
 	i = 0;
 	ptr = argv;
